Use null as a separator for xargs
This commit updates the invocation of `xargs` in scripts that send data
to logstash to add the `-0` flag which indicates a null byte is being
used as the separator between arguments. This separation is produced
by the use of `-print0` with `find` and GNU `xargs` on Linux will exit
with an error if the corresponding `-0` is not used.